move property stuff around
