+++ addon/properties/index.js
@@ -3,11 +3,16 @@ export {
 } from './-meta';
 
 export {
+  prop
+} from './-prop';
+
+export {
   first,
   find,
   paginated
 } from './proxy';
 
 export {
-  prop
-} from './prop';
+  model,
+  models
+} from './model';
