+++ addon/properties/model/index.js
@@ -1,0 +1,2 @@
+export { default as model } from './model';
+export { default as models } from './models';
