+++ addon/properties/proxy/index.js
@@ -1,0 +1,5 @@
+import proxy from './-create-proxy';
+
+export const first     = proxy('first');
+export const find      = proxy('find');
+export const paginated = proxy('paginated');
